+++ src/app/(route)/user/[userId]/_components/TabContents/TabContents.tsx
@@ -3,7 +3,7 @@
 import { PostListItem } from "@/components/domain";
 import { EmptyState, LoadingState } from "@/components/state";
 import { UserCommentsDataType, UserPostsDataType, UserProfileItem } from "@/api/fetch/user";
-import { UserProfileTabKey } from "../../_types/USER_TABS";
+import { UserProfileTabKey } from "../../_types/UserProfileTabType";
 import { UserCommentItem } from "../_internal";
 
 interface TabContentsProps {
