+++ src/app/(route)/user/[userId]/_components/USER_TABS.ts
@@ -3,5 +3,3 @@ export const USER_TABS = [
   { key: "comments", label: "댓글" },
   { key: "favorites", label: "즐겨찾기" },
 ] as const;
-
-export type UserProfileTabKey = (typeof USER_TABS)[number]["key"];
