+++ src/app/(route)/user/[userId]/_components/UserProfileView/UserProfileView.tsx
@@ -6,8 +6,8 @@ import { Tab } from "@/components/domain";
 import { useGetUserProfileById, UserTabType, UserUpperTabType } from "@/api/fetch/user";
 import UserHeader from "../UserHeader/UserHeader";
 import TabContents from "../TabContents/TabContents";
-import { USER_TABS } from "../../_types/USER_TABS";
-import { useUserProfileTabQuery } from "../../_hooks/useUserProfileTabQuery";
+import { USER_TABS } from "../USER_TABS";
+import { useUserProfileTabQuery } from "../../_hooks/useUserProfileTabQuery/useUserProfileTabQuery";
 import { useInfiniteScroll } from "@/hooks/useInfiniteScroll/useInfiniteScroll";
 
 const upperCase = (tab: UserTabType): UserUpperTabType => {
