+++ src/app/(route)/user/[userId]/_types/UserProfileTabType.ts
@@ -1,0 +1,3 @@
+import { USER_TABS } from "../_components/USER_TABS";
+
+export type UserProfileTabKey = (typeof USER_TABS)[number]["key"];
